kmod: Insmod using module name alone
The insmod utility in kmod currently inserts the
kernel module only using the path and extension given
as input, otherwise it fails to load the kernel module.

The change in this patch helps to insert the module
in the kernel using just the kernel name. The default
directory for locating the kernel modules is taken as
/lib/modules/`uname -r`.

The original behaviour where the modules can be inserted
using the absolute path and extension will still work
with this change.

Change-Id: I872fd432a7f9cac61ea2038c1dcebd5a90c25756
Signed-off-by: Sandhya KS <quic_sandhyks@quicinc.com>